requirements.txt
django
django-allauth
django-tastypie
django-guardian
django-sitetree
gunicorn
psycopg2
redis
six
south
sphinx
sphinx-rtd-theme==0.1.5

-e git+git://github.com/abourget/gevent-socketio.git#egg=gevent-socketio
